LF-4698 Small refactors, including timing of mock
Was sometimes getting no calls to axios on the first userRole only
